feat: Moves `ImageGallery` to fs/components
