+++ packages/components/src/organisms/ImageGallery/index.ts
@@ -1,4 +1,5 @@
+export { default } from './ImageGallery'
+export type { ImageElementData, ImageGalleryProps } from './ImageGallery'
+
 export { default as ImageGallerySelector } from './ImageGallerySelector'
-export { default as ImageGallery } from './ImageGallery'
 export { default as ImageZoom } from './ImageZoom'
-export type { ImageElementData } from './ImageGallery'
